add haproxy to ODCS compose configuration
Make the haproxy RPM available during the container build process so
that "microdnf install haproxy" succeeds.

In RHCS 5, we made the RPM available in our product composes (the
"Tools" repo) and passed the "yum_repourls" configuration into each
buildContainer Brew task.

In RHCS 6, we'll use ODCS composes for this RPM instead. We don't want
to ship our container-only packages to the Tools repository for
customers to access.  In other words, customers should not be able to
"yum install haproxy" on their own servers from
rhceph-6-tools-for-rhel-9-x86_64-rpms.